Environment
This creates the Environment class, along with the utilities to support
it.  The Environment class is a dictionary-like class extending the
SensitiveDict class from utils.py.  It provides access to environment
variables, including the ability to access list-like and set-like
variables as Python lists and sets, respectively; manipulation of the
working directory for executed processes through the 'cwd' property;
and a thin wrapper around subprocess.Popen() that makes use of the
environment and data.  Since Environment is a subclass of the
SensitiveDict, it is also possible to mask the values, allowing for
safe debugging that won't reveal the contents of sensitive variables.

For the Environment class, sensitive variables can be listed in the
special environment variable TIMID_SENSITIVE; any additional
sensitive variables set with declare_sensitive() will appear in
TIMID_SENSITIVE in subsequent calls to Environment.call().  Note
that order is not important (and thus not preserved) in the
TIMID_SENSITIVE environment variable.