Fix pep8 problem
The pep8 test didn't work properly, and I realized it was because of the
requirement for pwsafe, and how it was interpreted by setup.py, which
was installing timid into the virtual environment.  This fixes the
problem by using the same 'deps' setting as all other tests in tox.ini.