Steps
This creates the Step class, along with Action, Modifier, StepResult,
ConfigError, and StepAddress classes (among other internal ones).  These
classes together provide the ability to parse and represent a single
step in the test sequence, where a step consists of a single Action and
zero or more Modifier instances.

This also creates the first built-in Action, the "include" directive;
this directive allows one step file to reference another step file.