Add environment-related actions
This adds the environment-related actions 'env', 'chdir', and 'run'.
These actions alter the environment, alter the working directory, and
execute a command, respectively.  Note that the 'env' action is
actually implemented in abstract form in steps.py, as a very similar
one will be needed for manipulating variables for use with Jinja2
templates.