Allow BaseException through
Currently, all exceptions from extensions are ignored, even
BaseException.  This means that sys.exit() from an extension will not
work, and extensions are thus unable to stop timid execution even if
they are aware of a failure, e.g., an argument processing problem.
This change allows a BaseException to bubble up in the normal way.